Reorganize the directory structure
- Change `core.model` to `core.models`.
