+++ core/models/memory.py
