+++ core/processors/dialog/dialog_processors.py
@@ -9,7 +9,7 @@ from loguru._logger import Logger
 
 from core.strategy import DialogProcessStrategy
 from core.storage.db.base import Sqlstorage
-from core.model.app import AssistantRun
+from core.models.app import AssistantRun
 
 
 class OperationPriority(IntEnum):
