+++ core/storage/db/sqlite.py
@@ -15,8 +15,8 @@ from typing import Optional, List, Literal, Any, Dict
 from loguru import logger
 from sqlalchemy.schema import Table
 from tenacity import retry, stop_after_attempt, wait_exponential
-from core.model.memory import MemoryRow
-from core.model.app import AssistantRun
+from core.models.memory import MemoryRow
+from core.models.app import AssistantRun
 from core.storage.db.base import Sqlstorage
 from utils.log.logger_config import setup_logger
 from core.encryption import FernetEncryptor
